requirements.txt
aiohttp>=3.13.0 

PyMuPDF>=1.23.0

python-docx>=1.1.0

Pillow>=10.0.0

opencv-python>=4.8.0
numpy>=1.24.0

langchain-openai>=0.1.0
langchain>=0.1.0

pydantic>=2.0.0